separating execute and prepare
