+++ src/execute.rs
@@ -1,9 +1,6 @@
-use std::collections::HashSet;
-use std::env;
 use std::fs;
 use std::io::Error;
-use std::io::Write;
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 use std::process::Command;
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::mpsc::channel;
@@ -11,158 +8,54 @@ use std::sync::Arc;
 use std::thread::sleep;
 use std::time::{Duration, Instant};
 
-use ansi_term::Colour::{Cyan, Green, Yellow};
+use ansi_term::Colour::{Green, Yellow};
 use linked_hash_map::LinkedHashMap as Map;
 use notify::{raw_watcher, RawEvent, RecursiveMode, Watcher};
 
-use crate::commands::{Cmd, FileConfig, Repeat};
-use crate::consts::{CliArgs, BAR, DONKEY_COMMAND_ENV, DONKEY_DEPTH_ENV, DONKEY_FILE_ENV, DONKEY_KEEP_ENV, PATH_STR};
+use crate::commands::{Cmd, Repeat};
+use crate::consts::CliArgs;
 
-pub fn main(cmd_name: &str, config: &FileConfig, cmd: &Cmd, cli: &CliArgs, file_path: &PathBuf) -> Result<i32, String> {
-    let mut path_str: String = PATH_STR.to_string();
-    let mut run_depth: i32 = 0;
-    if let Ok(v) = env::var(DONKEY_DEPTH_ENV) {
-        path_str = format!("{}.{}", PATH_STR, v);
-        run_depth = v.parse::<i32>().unwrap_or(1);
-    }
-    let smart_prefix = match env::var(DONKEY_COMMAND_ENV) {
-        Ok(c) => format!("{} > {}", c, cmd_name),
-        _ => cmd_name.to_string(),
-    };
-    let mut args: Vec<String> = vec![path_str.clone()];
-    args.extend(cmd.args.iter().cloned());
-    args.extend(cli.args.iter().cloned());
+pub struct Run {
+    pub cmd_name: String,
+    pub args: Vec<String>,
+    pub env: Map<String, String>,
+    pub working_dir: PathBuf,
+    pub tmp_path: PathBuf,
+    pub print_summary: bool,
+}
 
-    let mut env: StrMap = Map::new();
-    merge_maps(&mut env, &config.env);
-    merge_maps(&mut env, &cmd.env);
-    env.insert(DONKEY_DEPTH_ENV.to_string(), (run_depth + 1).to_string());
-    env.insert(DONKEY_FILE_ENV.to_string(), full_path(file_path));
-    env.insert(DONKEY_COMMAND_ENV.to_string(), smart_prefix.clone());
-    env.insert(
-        DONKEY_KEEP_ENV.to_string(),
-        String::from(if cli.keep_tmp { "1" } else { "0" }),
-    );
-
-    let working_dir = get_working_dir(&cmd, file_path)?;
-    let path = working_dir.join(&path_str);
-    write(cmd_name, &path, cmd, &args, &env, config, smart_prefix)?;
-
-    let print_summary: bool = run_depth == 0;
-    if print_summary {
-        eprintlnc!(
-            Green,
-            r#"Running command "{}" from {}..."#,
-            cmd_name,
-            file_path.display()
-        );
-    }
-
+pub fn main(run: &Run, cmd: &Cmd, cli: &CliArgs) -> Result<i32, String> {
     let exit_code = match &cmd.repeat {
-        Some(Repeat::Periodic { interval: i }) => {
-            run_command_periodic(*i, cmd_name, cmd, &args, &env, working_dir, print_summary).map_err(error_str)
-        }
-        Some(Repeat::Watch { debounce: d, dir }) => {
-            run_command_watch(*d, dir, cmd_name, cmd, &args, &env, working_dir, print_summary)
-        }
-        None => run_command_once(cmd_name, cmd, &args, &env, &working_dir, print_summary).map_err(error_str),
+        Some(Repeat::Periodic { interval: i }) => run_command_periodic(&run, &cmd, *i).map_err(error_str),
+        Some(Repeat::Watch { debounce: d, dir }) => run_command_watch(&run, &cmd, *d, dir),
+        None => run_command_once(&run, &cmd).map_err(error_str),
     };
-    delete(&path, cli.keep_tmp)?;
+    delete(&run.tmp_path, cli.keep_tmp)?;
     match exit_code {
         Ok(t) => Ok(t),
         Err(e) => err!(
             "failed to execute command \"{} {}\": {}",
             cmd.executable(),
-            args.join(" "),
+            run.args.join(" "),
             e
         ),
     }
 }
 
-fn write(
-    cmd_name: &str,
-    path: &PathBuf,
-    cmd: &Cmd,
-    args: &[String],
-    env: &StrMap,
-    config: &FileConfig,
-    smart_prefix: String,
-) -> Result<(), String> {
-    if path.exists() {
-        return err!(
-            "Error writing temporary file:\n  {} already exists, donkey-make may be running already",
-            path.display()
-        );
-    }
-
-    let prefix: Vec<String> = vec![
-        String::from(BAR),
-        format!(
-            "This is a temporary file generated by donkey-make to execute the command: \"{}\"",
-            cmd_name
-        ),
-        format!("Command to be executed: \"{} {}\"", cmd.executable(), args.join(" ")),
-        String::from("Environment variables set:"),
-        format!("{:?}", env),
-        String::from("This file should only exist very temporarily while it's be executed."),
-        String::from(BAR),
-    ];
-    let comment = if cmd.executable().starts_with("node") {
-        "//"
-    } else {
-        "#"
-    };
-    let sep = format!("\n{} ", comment);
-
-    let script: String = if cmd.smart() {
-        let donk_exe = match env::current_exe() {
-            Ok(ex) => full_path(&ex),
-            Err(e) => return err!("finding current executable for smart script failed: {}", e),
-        };
-        let mut cmd_tree: HashSet<String> = HashSet::new();
-        cmd_tree.insert((*cmd_name).to_string());
-        build_smart_script(&cmd, smart_prefix, &donk_exe, config, &mut cmd_tree)?
-    } else {
-        cmd.run.join("\n")
-    };
-
-    let content = format!("{} {}\n{}", comment, prefix.join(&sep), script);
-
-    match create_file(path, &content) {
-        Ok(_) => Ok(()),
-        Err(e) => err!("Error writing temporary file {}:\n  {}", path.display(), e),
-    }
-}
-
-fn run_command_once(
-    cmd_name: &str,
-    cmd: &Cmd,
-    args: &[String],
-    env: &StrMap,
-    working_dir: &PathBuf,
-    print_summary: bool,
-) -> Result<i32, Error> {
+fn run_command_once(run: &Run, cmd: &Cmd) -> Result<i32, Error> {
     let sig = register_signals()?;
-    run_command(cmd_name, cmd, &args, &env, &working_dir, print_summary, &sig)
+    run_command(run, cmd, &sig)
 }
 
 const WAIT_MS: u64 = 20;
 
-fn run_command_periodic(
-    interval: f32,
-    cmd_name: &str,
-    cmd: &Cmd,
-    args: &[String],
-    env: &StrMap,
-    working_dir: PathBuf,
-    print_summary: bool,
-) -> Result<i32, Error> {
+fn run_command_periodic(run: &Run, cmd: &Cmd, interval: f32) -> Result<i32, Error> {
     let sleep_steps = (interval * 1000.0 / (WAIT_MS as f32)) as u64;
     let sleep_time = Duration::from_millis(WAIT_MS);
     let sig = register_signals()?;
 
     loop {
-        let status_code = run_command(cmd_name, cmd, &args, &env, &working_dir, print_summary, &sig)?;
+        let status_code = run_command(run, cmd, &sig)?;
         if status_code != 0 {
             return Ok(status_code);
         }
@@ -175,16 +68,7 @@ fn run_command_periodic(
     }
 }
 
-fn run_command_watch(
-    debounce: f32,
-    watch_dir: &str,
-    cmd_name: &str,
-    cmd: &Cmd,
-    args: &[String],
-    env: &StrMap,
-    working_dir: PathBuf,
-    print_summary: bool,
-) -> Result<i32, String> {
+fn run_command_watch(run: &Run, cmd: &Cmd, debounce: f32, watch_dir: &str) -> Result<i32, String> {
     let debounce_dur = Duration::from_millis((debounce * 1000.0) as u64);
     let recv_timeout = Duration::from_millis(WAIT_MS);
 
@@ -215,8 +99,7 @@ fn run_command_watch(
         println!("event: {:?}", events);
         first_event = None;
         events.clear();
-        let status_code =
-            run_command(cmd_name, cmd, &args, &env, &working_dir, print_summary, &sig).map_err(error_str)?;
+        let status_code = run_command(run, cmd, &sig).map_err(error_str)?;
         if status_code != 0 {
             return Ok(status_code);
         }
@@ -226,17 +109,9 @@ fn run_command_watch(
     }
 }
 
-fn run_command(
-    cmd_name: &str,
-    cmd: &Cmd,
-    args: &[String],
-    env: &StrMap,
-    working_dir: &PathBuf,
-    print_summary: bool,
-    sig: &Signal,
-) -> Result<i32, Error> {
-    let mut c = Command::new(&cmd.executable());
-    c.args(args).envs(env).current_dir(working_dir);
+fn run_command(run: &Run, cmd: &Cmd, sig: &Signal) -> Result<i32, Error> {
+    let mut c = Command::new(cmd.executable());
+    c.args(&run.args).envs(&run.env).current_dir(&run.working_dir);
 
     let start = Instant::now();
     let status = c.status()?;
@@ -244,14 +119,14 @@ fn run_command(
     let dur_str = format_duration(duration);
     if let Some(c) = status.code() {
         if c == 0 {
-            if print_summary {
-                eprintlnc!(Green, "Command \"{}\" successful in {} 👍", cmd_name, dur_str);
+            if run.print_summary {
+                eprintlnc!(Green, "Command \"{}\" successful in {} 👍", run.cmd_name, dur_str);
             }
         } else {
             eprintlnc!(
                 Yellow,
                 "Command \"{}\" failed in {}, exit code {} 👎",
-                cmd_name,
+                run.cmd_name,
                 dur_str,
                 c
             );
@@ -261,7 +136,7 @@ fn run_command(
         eprintlnc!(
             Yellow,
             "Command \"{}\" kill with signal {} after {} 👎",
-            cmd_name,
+            run.cmd_name,
             signal_name(sig).unwrap_or("UNKNOWN"),
             dur_str
         );
@@ -278,12 +153,6 @@ fn delete(path: &PathBuf, keep: bool) -> Result<(), String> {
             }
         };
     }
-    Ok(())
-}
-
-fn create_file(path: &Path, content: &str) -> std::io::Result<()> {
-    let mut f = fs::File::create(path)?;
-    f.write_all(content.as_bytes())?;
     Ok(())
 }
 
@@ -313,87 +182,11 @@ fn signal_name(sig: &Signal) -> Option<&'static str> {
     }
 }
 
-const NO_ECHO_PREFIX: char = '_';
-const DONK_PREFIX: char = '+';
-const INLINE_PREFIX: char = '<';
-const PREFIXES: [char; 3] = [NO_ECHO_PREFIX, DONK_PREFIX, INLINE_PREFIX];
-
-fn build_smart_script(
-    cmd: &Cmd,
-    smart_prefix: String,
-    donk_exe: &str,
-    config: &FileConfig,
-    cmd_tree: &mut HashSet<String>,
-) -> Result<String, String> {
-    let all = cmd.run.join("\n");
-    let lines: Vec<&str> = all.split('\n').collect();
-    let len = lines.len();
-
-    let mut script: Vec<String> = vec!["set -e".to_string()];
-    for line in lines {
-        if !PREFIXES.iter().any(|&prefix| line.starts_with(prefix)) {
-            let coloured = epaint!(Cyan, format!("{} > {}", smart_prefix, line));
-            script.push(format!(">&2 echo '{}'", coloured));
-        }
-
-        let mut ex_line = if line.starts_with(NO_ECHO_PREFIX) {
-            line[1..].to_string()
-        } else {
-            line.to_string()
-        };
-
-        if ex_line.starts_with(INLINE_PREFIX) {
-            let sub_cmd_name = &ex_line[1..].trim().to_string();
-            if cmd_tree.contains(sub_cmd_name) {
-                return err!(
-                    "Command \"{}\" reused in an inline sub-command, this would cause infinite recursion",
-                    sub_cmd_name
-                );
-            }
-            cmd_tree.insert(sub_cmd_name.clone().to_string());
-            let sub_cmd = get_sub_command(config, sub_cmd_name)?;
-            let sub_cmd_prefix = format!("{} > {}", smart_prefix, sub_cmd_name);
-            ex_line = build_smart_script(sub_cmd, sub_cmd_prefix, donk_exe, config, &mut *cmd_tree)?;
-        } else {
-            if len == 1 && !line.contains('$') {
-                // must be the first line
-                ex_line = format!("{} $@", line)
-            }
-            if ex_line.starts_with(DONK_PREFIX) {
-                ex_line = format!("{} {}", donk_exe, &ex_line[1..]);
-            }
-        }
-        script.push(ex_line);
-    }
-    Ok(script.join("\n"))
-}
-
-fn get_sub_command<'a>(config: &'a FileConfig, cmd_name: &str) -> Result<&'a Cmd, String> {
-    match config.commands.get(cmd_name) {
-        Some(c) => {
-            if c.smart() {
-                Ok(c)
-            } else {
-                err!(
-                    "Sub-command \"{}\" not a bash-smart script, remove \"ex:\" or use '{}' not '{}'",
-                    cmd_name,
-                    DONK_PREFIX,
-                    INLINE_PREFIX
-                )
-            }
-        }
-        None => err!(
-            "Sub-command \"{}\" not found, commands available are:\n  {}",
-            cmd_name,
-            config.keys().join(", ")
-        ),
-    }
-}
-
-type StrMap = Map<String, String>;
-
-fn merge_maps(base: &mut StrMap, update: &StrMap) {
-    base.extend(update.iter().map(|(k, v)| (k.clone(), v.clone())));
+fn error_str<T>(e: T) -> String
+where
+    T: std::fmt::Display,
+{
+    format!("{}", e)
 }
 
 fn format_duration(duration: Duration) -> String {
@@ -405,47 +198,6 @@ fn format_duration(duration: Duration) -> String {
         }
         d => format!("{}s", d.as_secs()),
     }
-}
-
-fn full_path(path: &PathBuf) -> String {
-    match path.canonicalize() {
-        Ok(p) => p.to_string_lossy().to_string(),
-        _ => path.to_string_lossy().to_string(),
-    }
-}
-
-fn get_working_dir(cmd: &Cmd, file_path: &PathBuf) -> Result<PathBuf, String> {
-    match &cmd.working_dir {
-        Some(wd) => {
-            let mut path = PathBuf::from(&wd);
-            if path.is_relative() {
-                let file_dir = match file_path.parent() {
-                    Some(p) => p,
-                    _ => return err!("file path appears to have no parent directory"),
-                };
-                path = file_dir.join(&wd).to_path_buf();
-            }
-            if !path.is_dir() {
-                err!("\"{}\" is not a directory", &wd)
-            } else {
-                Ok(match path.canonicalize() {
-                    Ok(p) => p,
-                    _ => path,
-                })
-            }
-        }
-        _ => match env::current_dir() {
-            Ok(p) => Ok(p),
-            Err(e) => err!("unable to resolve current working directory: {}", e),
-        },
-    }
-}
-
-fn error_str<T>(e: T) -> String
-where
-    T: std::fmt::Display,
-{
-    format!("{}", e)
 }
 
 #[cfg(test)]
